feat(coin-hunting): Enhance map markers functionality with loading and error states
- Updated CoinHuntingScreen to pass collectedIds, loadingMarkers, and errorMarkers to MapMarkers.
- Refactored MapMarkers to accept new props and removed internal state management for markers.
- Implemented fetching of markers and collectedIds in useCoinHunting hook, including loading and error handling.
